Initial Project Upload
Created the unity project and set the appropriate setting. Please clone and update your local repositories ASAP. We will build on this as our tutorial so everyone gets familiar with Unity and we can begin our actual game later on. You should be able to open the game now with your Unity IDE.